+++ packages/plugin-core/src/WorkspaceWatcher.ts
@@ -1,6 +1,5 @@
 import { NoteProps, NoteUtils, Time, VaultUtils } from "@dendronhq/common-all";
-import { genHash } from "@dendronhq/common-server";
-import { HistoryService } from "@dendronhq/engine-server";
+import { HistoryService, WorkspaceUtils } from "@dendronhq/engine-server";
 import _ from "lodash";
 import path from "path";
 import {
@@ -11,6 +10,7 @@ import {
   TextEdit,
   window,
   workspace,
+  TextDocument,
 } from "vscode";
 import { Logger } from "./logger";
 import { DendronWorkspace, getWS } from "./workspace";
@@ -19,12 +19,17 @@ export class WorkspaceWatcher {
   activate(context: ExtensionContext) {
     workspace.onWillSaveTextDocument(
       this.onWillSaveTextDocument,
-      null,
+      this,
       context.subscriptions
     );
     workspace.onDidChangeTextDocument(
       this.onDidChangeTextDocument,
-      null,
+      this,
+      context.subscriptions
+    );
+    workspace.onDidOpenTextDocument(
+      this.onDidOpenTextDocument,
+      this,
       context.subscriptions
     );
   }
@@ -82,9 +87,9 @@ export class WorkspaceWatcher {
       return { changes: [] };
     }
     const match = NoteUtils.RE_FM_UPDATED.exec(content);
-    const noteHash = genHash(content);
     let changes: TextEdit[] = [];
-    if (match && note.contentHash && note.contentHash !== noteHash) {
+
+    if (match && WorkspaceUtils.noteContentChanged({ content, note })) {
       Logger.info({ ctx, match, msg: "update activeText editor" });
       const startPos = ev.document.positionAt(match.index);
       const endPos = ev.document.positionAt(match.index + match[0].length);
@@ -99,5 +104,31 @@ export class WorkspaceWatcher {
       ev.waitUntil(p);
     }
     return { changes };
+  }
+
+  onDidOpenTextDocument(document: TextDocument) {
+    this._openedDocuments.set(document.uri.fsPath, document);
+  }
+
+  /** The documents that have been opened during this session that have not been viewed yet in the editor. */
+  private _openedDocuments: Map<string, TextDocument> = new Map();
+
+  /** Do not use this function, please go to `WindowWatcher.onFirstOpen() instead.`
+   *
+   * Checks if the given document has been opened for the first time during this session, and marks the document as being processed.
+   *
+   * Certain actions (such as folding and adjusting the cursor) need to be done only the first time a document is opened.
+   * While the `WorkspaceWatcher` sees when new documents are opened, the `TextEditor` is not active at that point, and we can not
+   * perform these actions. This code allows `WindowWatcher` to check when an editor becomes active whether that editor belongs to an newly opened document.
+   *
+   * Mind that this method is not idempotent, checking the same document twice will always return false for the second time.
+   */
+  public getNewlyOpenedDocument(document: TextDocument): boolean {
+    const key = document.uri.fsPath;
+    if (this._openedDocuments?.has(key)) {
+      this._openedDocuments.delete(key);
+      return true;
+    }
+    return false;
   }
 }