+++ packages/plugin-core/src/test/suite-integ/FileWatcher.test.ts
@@ -1,17 +1,18 @@
 import { NoteUtils } from "@dendronhq/common-all";
 import { AssertUtils, NoteTestUtilsV4 } from "@dendronhq/common-test-utils";
 import { ENGINE_HOOKS_MULTI } from "@dendronhq/engine-test-utils";
+import _ from "lodash";
 import { describe } from "mocha";
 import path from "path";
 import * as vscode from "vscode";
-import { VaultWatcher } from "../../fileWatcher";
+import { FileWatcher } from "../../fileWatcher";
 import { VSCodeUtils } from "../../utils";
 import { expect } from "../testUtilsv2";
 import { runLegacyMultiWorkspaceTest, setupBeforeAfter } from "../testUtilsV3";
 
 suite("FileWatcher", function () {
   let ctx: vscode.ExtensionContext;
-  let watcher: VaultWatcher;
+  let watcher: FileWatcher;
 
   ctx = setupBeforeAfter(this, {
     beforeHook: () => {},
@@ -29,7 +30,7 @@ suite("FileWatcher", function () {
             vault: vaults[0],
             wsRoot,
           });
-          watcher = new VaultWatcher({
+          watcher = new FileWatcher({
             wsRoot,
             vaults,
           });
@@ -50,7 +51,7 @@ suite("FileWatcher", function () {
       });
     });
 
-    test("onDidChange", function (done) {
+    test("onDidChange: change", function (done) {
       runLegacyMultiWorkspaceTest({
         ctx,
         postSetupHook: ENGINE_HOOKS_MULTI.setupBasicMulti,
@@ -62,18 +63,38 @@ suite("FileWatcher", function () {
             const selection = new vscode.Selection(pos, pos);
             builder.replace(selection, `Hello`);
           });
-          watcher = new VaultWatcher({
+          watcher = new FileWatcher({
             wsRoot,
             vaults,
           });
           const uri = editor.document.uri;
-          await watcher.onDidChange(uri);
+          const resp = await watcher.onDidChange(uri);
+          expect(resp?.contentHash).toEqual("465a4f4ebf83fbea836eb7b8e8e040ec");
           expect(
             await AssertUtils.assertInString({
               body: engine.notes["foo"].body,
               match: ["Hello"],
             })
           ).toBeTruthy();
+          done();
+        },
+      });
+    });
+
+    test("onDidChange: no change", function (done) {
+      runLegacyMultiWorkspaceTest({
+        ctx,
+        postSetupHook: ENGINE_HOOKS_MULTI.setupBasicMulti,
+        onInit: async ({ vaults, wsRoot, engine }) => {
+          const foo = engine.notes["foo"];
+          const editor = await VSCodeUtils.openNote(foo);
+          watcher = new FileWatcher({
+            wsRoot,
+            vaults,
+          });
+          const uri = editor.document.uri;
+          const resp = await watcher.onDidChange(uri);
+          expect(_.isUndefined(resp)).toBeTruthy();
           done();
         },
       });