+++ packages/pods-core/src/builtin/GraphvizPod.ts
@@ -2,8 +2,9 @@ import fs from "fs-extra";
 import _ from "lodash";
 import { DLink, NoteProps, NoteUtils } from "@dendronhq/common-all";
 import path from "path";
-import { ExportPod, ExportPodPlantOpts } from "../basev3";
-import { ExportPodConfig } from "../basev3";
+import { ExportPod, ExportPodPlantOpts, ExportPodConfig } from "../basev3";
+import { JSONSchemaType } from "ajv";
+import { PodUtils } from "../utils";
 
 const ID = "dendron.graphviz";
 
@@ -31,23 +32,24 @@ export class GraphvizExportPod extends ExportPod<GraphvizExportConfig> {
   static id: string = ID;
   static description: string = "export notes in Graphviz DOT format";
 
-  get config() {
-    return super.config.concat([
-      {
-        key: "showGraphByHierarchy",
-        description:
-          "Include hierarchical note connections (e.g. parent -> child connections)",
-        default: true,
-        type: "boolean",
+  get config(): JSONSchemaType<GraphvizExportConfig> {
+    return PodUtils.createExportConfig({
+      required: [],
+      properties: {
+        showGraphByHierarchy: {
+          type: "boolean",
+          description:
+            "Include hierarchical note connections (e.g. parent -> child connections)",
+          default: true,
+        },
+        showGraphByEdges: {
+          type: "boolean",
+          description:
+            "Include linked note relationships, e.g. note with [[link]] -> another note",
+          default: false,
+        },
       },
-      {
-        key: "showGraphByEdges",
-        description:
-          "Include linked note relationships, e.g. note with [[link]] -> another note",
-        default: false,
-        type: "boolean",
-      },
-    ]);
+    }) as JSONSchemaType<GraphvizExportConfig>;
   }
 
   // Dashes are not allowed, so they are removed.